+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Create `FlexCarousel` component.
+- Feature: Create `FlexCarousel` and `FlexCarouselInner` components.
 - Feature: Add `hidePaginationOnMobile` prop to `Carousel`.
 - Fix: Remove `hover` state on `Select-control` class.
 - Fix: Update size on `ButtonQuestionMarkIcon` component.